multi segmented pool
